Updated extended_wikidata_search function implementation
