+++ sand/extensions/search/wikidata_search.py
@@ -5,6 +5,8 @@ from sand.extension_interface.search import IEntitySearch, IOntologySearch
 from sand.models.entity import Entity
 from sand.models.ontology import OntClass, OntProperty, OntClassAR
 from sand.models.search import SearchResult
+from sand.extensions.search.default_search import DefaultSearch
+from sand.extensions.search.aggregated_search import AggregatedSearch
 
 
 class WikidataSearch(IEntitySearch, IOntologySearch):
@@ -22,6 +24,13 @@ class WikidataSearch(IEntitySearch, IOntologySearch):
             "srprop": "snippet|titlesnippet"
         }
         self.ont_class_ar = None
+
+    def extended_wikidata_search(self) -> Union[IEntitySearch,IOntologySearch]:
+        """extended version of wikidata search by aggregating default search"""
+        search = AggregatedSearch()
+        search.add(DefaultSearch())
+        search.add(WikidataSearch())
+        return search
 
     def get_class_search_params(self, search_text: str) -> Dict:
         """Updates class search parameters for wikidata API"""
